refactor(authority): bring back NDD to authority crate
