+++ crates/consensus/authority/src/comet_bft/mod.rs
@@ -1,2 +1,3 @@
 /// ABCI client implementation and consensus engine
 pub mod abci;
+mod proto_debug;
